echo/pyproject.toml
[project]
name = "echo"
version = "0.1.0"
description = "Tweet discovery and engagement system"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "asyncpg>=0.30.0",
    "mcp>=1.26.0",
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
testpaths = ["compose/tests"]
